Changed package structure. Added config files and monoDB templates. Allowed 2 repo's to store into the same MongoDb Atlas cluster.
