+++ cst438-project2/src/main/java/com/groupO/project2/MultipleMongoProperties.java
@@ -1,0 +1,14 @@
+package com.groupO.project2;
+
+import org.springframework.boot.autoconfigure.mongo.MongoProperties;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+import lombok.Data;
+
+@Data
+@ConfigurationProperties(prefix = "mongodb")
+public class MultipleMongoProperties {
+
+    private MongoProperties primary = new MongoProperties();
+    private MongoProperties secondary = new MongoProperties();
+}
