+++ cst438-project2/src/main/java/com/groupO/project2/UserLoader.java
@@ -4,6 +4,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.stereotype.Component;
 
+import com.groupO.project2.repository.userRepo.User;
+import com.groupO.project2.repository.userRepo.UserRepository;
+
 @Component 
 public class UserLoader implements CommandLineRunner {
