--- cst438-project2/src/main/java/com/groupO/project2/UserRepository.java
@@ -1,12 +1,0 @@
-package com.groupO.project2;
-import java.util.List;
-
-import org.springframework.data.mongodb.repository.MongoRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends MongoRepository<User, String> {
-
-  public User findByEmail(String email);
-  //public User findById(String id);
-}
