+++ cst438-project2/src/main/java/com/groupO/project2/configurations/NewDb2Config.java
@@ -1,0 +1,13 @@
+package com.groupO.project2.configurations;
+
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.mongodb.repository.config.EnableMongoRepositories;
+
+@Configuration
+@EnableMongoRepositories(basePackages = {"com.groupO.project2.repository.userRepo"},
+mongoTemplateRef = NewDb2Config.MONGO_TEMPLATE
+)
+public class NewDb2Config {
+    protected static final String MONGO_TEMPLATE = "newdb2MongoTemplate";
+}
