+++ cst438-project2/src/main/java/com/groupO/project2/repository/itemRepo/ItemRepository.java
@@ -1,0 +1,10 @@
+package com.groupO.project2.repository.itemRepo;
+
+import org.springframework.data.mongodb.repository.MongoRepository;
+import org.springframework.stereotype.Repository;
+
+public interface ItemRepository extends MongoRepository<Item, String> {
+
+  public Item findByItemLink(String link);
+  //public Item findById(String id);
+}
