+++ cst438-project2/src/main/java/com/groupO/project2/repository/userRepo/UserRepository.java
@@ -1,0 +1,13 @@
+package com.groupO.project2.repository.userRepo;
+import java.util.List;
+
+import org.springframework.data.mongodb.repository.MongoRepository;
+import org.springframework.stereotype.Repository;
+
+public interface UserRepository extends MongoRepository<User, String> {
+
+  public User findByEmail(String email);
+  //public User findById(String id);
+}
+
+
