+++ cst438-project2/src/main/resources/application.properties
@@ -1,9 +1,22 @@
 # spring.batch.job.enabled=false
 spring.data.rest.base-path=/api
 #spring.main.allow-bean-definition-overriding=true
-spring.data.mongodb.uri=mongodb+srv://admin:admin@Project2.i2sye.mongodb.net/Project2?retryWrites=true&w=majority
-spring.data.mongodb.database=Project2
+#spring.data.mongodb.uri=mongodb+srv://admin:admin@Project2.i2sye.mongodb.net/Project2?retryWrites=true&w=majority
+#spring.data.mongodb.database=Project2
+
+spring.autoconfigure.exclude= org.springframework.boot.autoconfigure.mongo.MongoAutoConfiguration
+spring.data.mongodb.newdb1.uri=mongodb+srv://admin:admin@Project2.i2sye.mongodb.net/Project2?retryWrites=true&w=majority
+spring.data.mongodb.newdb2.uri=mongodb+srv://admin:admin@Project2.i2sye.mongodb.net/Project2?retryWrites=true&w=majority
 
 #spring:
 #  autoconfigure:
 #    exclude: org.springframework.boot.autoconfigure.mongo.MongoAutoConfiguration
+#    
+#mongodb:
+#  primary:
+#    uri: mongodb+srv://admin:admin@Project2.i2sye.mongodb.net/Project2?retryWrites=true&w=majority
+#    database: Project2
+#  secondary:
+#    uri: mongodb+srv://admin:admin@Project2.i2sye.mongodb.net/Project2?retryWrites=true&w=majority
+#    database: Project2
+
